requirements.txt
requests==2.32.3
playwright==1.51.0
pytest>=7.0.0
python-dotenv>=1.0.0

# StandX Exchange dependencies
cryptography>=41.0.0
base58>=2.1.0
web3>=6.0.0
eth-account>=0.8.0
pyyaml>=6.0.0

# Technical Analysis dependencies
pandas>=2.0.0
TA-Lib>=0.4.28
